isolate getRoute function under utils
