+++ ts/store/reducers/debug.ts
@@ -1,5 +1,4 @@
 import { getType } from "typesafe-actions";
-
 import { setDebugModeEnabled } from "../actions/debug";
 import { Action } from "../actions/types";
 import { GlobalState } from "./types";
